move deeplink and other logic from decider to app_shell
